+++ NEWS.md
@@ -176,6 +176,9 @@ Library improvements
   * `@test isequal(x, y)` and `@test isapprox(x, y)` now prints an evaluated expression when
     the test fails ([#22296]).
 
+  * `isapprox(x,y)` now tests `norm(x-y) <= max(atol, rtol*max(norm(x), norm(y)))`
+    rather than `norm(x-y) <= atol + ...`.
+
   * Uses of `Val{c}` in `Base` has been replaced with `Val{c}()`, which is now easily
     accessible via the `@pure` constructor `Val(c)`. Functions are defined as
     `f(::Val{c}) = ...` and called by `f(Val(c))`. Notable affected functions include: